chore: add seed for products,  brands and categories

+++ database/migrations/2025_03_14_104846_create_products_table.php
@@ -6,28 +6,28 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
-    public function up(): void
-    {
-        Schema::create('products', function (Blueprint $table) {
-            $table->uuid('id')->primary();
-            $table->string('name')->unique();
-            $table->string('description')->nullable();
-            $table->decimal('price', 10, 2);
-            $table->integer('stock');
-            $table->foreignUuid('brand_id')->constrained('brands');
-            $table->foreignUuid('category_id')->constrained('categories');
-            $table->timestamps();
-        });
-    }
+  /**
+   * Run the migrations.
+   */
+  public function up(): void
+  {
+    Schema::create('products', function (Blueprint $table) {
+      $table->uuid('id')->primary();
+      $table->string('name')->unique();
+      $table->string('description')->nullable();
+      $table->bigInteger('price');
+      $table->integer('stock');
+      $table->foreignUuid('brand_id')->constrained('brands');
+      $table->foreignUuid('category_id')->constrained('categories');
+      $table->timestamps();
+    });
+  }
 
-    /**
-     * Reverse the migrations.
-     */
-    public function down(): void
-    {
-        Schema::dropIfExists('products');
-    }
+  /**
+   * Reverse the migrations.
+   */
+  public function down(): void
+  {
+    Schema::dropIfExists('products');
+  }
 };
